requirements.txt
Flask==3.1.0
supabase==2.15.0
opencv-python-headless==4.11.0.86
mediapipe==0.10.21
numpy==1.26.4
tensorflow==2.19.0
tflite-runtime==2.19.0
python-dotenv==1.1.0
gunicorn>=20.0 # For production deployment
